// src/features/instagram/types.ts
export type InstagramAnalysis = {
    sourceName: string
    followingCount: number
    followersCount: number
    mutualCount: number
    notFollowingBackCount: number
    fansCount: number
    following: string[]
    followers: string[]
    mutuals: string[]
    notFollowingBack: string[]
    fans: string[]
}